Deduplicate monixAsyncQueue
