+++ async-http-client-backend/monix/src/main/scala/sttp/client/asynchttpclient/monix/MonixWebSocketHandler.scala
@@ -1,28 +1,13 @@
 package sttp.client.asynchttpclient.monix
 
 import monix.eval.Task
-import monix.execution.{Scheduler, AsyncQueue => MonixAsyncQueue}
+import monix.execution.Scheduler
 import sttp.client.asynchttpclient.WebSocketHandler
-import sttp.client.asynchttpclient.internal.{AsyncQueue, NativeWebSocketHandler}
-import sttp.client.impl.monix.TaskMonadAsyncError
+import sttp.client.asynchttpclient.internal.NativeWebSocketHandler
+import sttp.client.impl.monix.{MonixAsyncQueue, TaskMonadAsyncError}
 import sttp.client.ws.WebSocket
-import sttp.model.ws.WebSocketBufferFull
 
 object MonixWebSocketHandler {
-  private class MonixAsyncQueue[A](bufferCapacity: Option[Int])(implicit s: Scheduler) extends AsyncQueue[Task, A] {
-    private val queue = bufferCapacity match {
-      case Some(capacity) => MonixAsyncQueue.bounded[A](capacity)
-      case None           => MonixAsyncQueue.unbounded[A]()
-    }
-
-    override def clear(): Unit = queue.clear()
-    override def offer(t: A): Unit = {
-      if (!queue.tryOffer(t)) {
-        throw new WebSocketBufferFull()
-      }
-    }
-    override def poll: Task[A] = Task.deferFuture(queue.poll())
-  }
 
   /**
     * Creates a new [[WebSocketHandler]] which should be used *once* to send and receive from a single websocket.
