--- httpclient-backend/monix/src/main/scala/sttp/client/httpclient/monix/internal/AsyncQueue.scala
@@ -1,19 +1,0 @@
-package sttp.client.httpclient.monix.internal
-
-trait AsyncQueue[F[_], T] {
-
-  /**
-    * Eagerly clears the queue.
-    */
-  def clear(): Unit
-
-  /**
-    * Eagerly adds the given item to the queue.
-    */
-  def offer(t: T): Unit
-
-  /**
-    * Takes an element from the queue or suspends, until one is available. May be eager or lazy, depending on `F`.
-    */
-  def poll: F[T]
-}
